Support for multi periods in remote entity.
The earlier version does not support multi periods in remote entity. The main issue being that the remote entity is not a proper XML document.
In this commit, I manually first add a root element around the remote entity to first ensure that it is a XML parsable document. Further, each of the multi periods have to appropriately passed and checked. This is also done.